requirements.txt
web3==4.9.2
py-solc==3.2.0
rlp
eth_utils
two1
requests